fixup! feat: share useful tools from a3p-integration
